setup react router (dom) and plug newly created pages on it
