+++ front-end/src/types.ts
@@ -1,15 +1,15 @@
 export interface Review {
-    id: string;
-    rating: number;
-    author: {
-      name: string;
-    };
-  }
+  id: string;
+  rating: number;
+  content: string; // ✅ this was missing
+  author: {
+    name: string;
+  };
+}
   
-  export interface Game {
-    id: string;
-    title: string;
-    platform: string[];
-    reviews: Review[];
-  }
-  
+export interface Game {
+  id: string;
+  title: string;
+  platform: string[];
+  reviews: Review[];
+}
